requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Security-Too==5.3.3
Werkzeug==3.0.1
email-validator==2.1.0
bcrypt==4.1.2
